TFM/cloud-functions/cv-uploader/requirements.txt
google-cloud-storage
google-cloud-bigquery
google-cloud-logging
functions-framework 